1.2.0 update

+++ loadwatch.spec
@@ -1,11 +1,16 @@
-%global log_dir /var/log/loadwatch
+%global log_dir var/log/loadwatch
+%global legacy_log_dir root/loadwatch
+%global conf etc/default/%{name}
+%global bin usr/local/lp/bin/%{name}
+%global cron etc/cron.d/%{name}.cron
 Summary: A script to monitor a system for abnormal conditions, and log data
 Name: loadwatch
-Version: 1.1.0
+Version: 1.2.0
 Release: 0
 URL: https://github.com/jakdept/loadwatch
 License: MIT
 Group: Applications/System
+Packager: Jack Hayhurst <support-tools@liquidweb.com>
 BuildRoot: %{_topdir}/%{name}
 BuildArch: noarch
 Requires: bash, cronie, lynx, sed, /bin/find
@@ -19,7 +24,6 @@ for later inspection.
 
 %prep
 
-
 %build
 
 %install
@@ -28,24 +32,41 @@ mkdir -p \
   %{buildroot}/etc/default \
   %{buildroot}/etc/cron.d \
   %{buildroot}/root \
-  %{buildroot}/var/log/loadwatch
+  %{buildroot}/%{log_dir}
 echo %{buildroot}
 echo %{_sourcedir}/%{name}
-ln -s -f -L /var/log/loadwatch %{buildroot}/root/loadwatch
-install -m 0700 %{_sourcedir}/%{name}/loadwatch %{buildroot}/usr/local/lp/bin/loadwatch
-install -m 755 %{_sourcedir}/%{name}/loadwatch.env %{buildroot}/etc/default/loadwatch
-install -m 0600 %{_sourcedir}/%{name}/loadwatch.cron %{buildroot}/etc/cron.d/loadwatch.cron
+ln -s -f -L /%{log_dir} %{buildroot}/%{legacy_log_dir}
+install -m 0700 %{_sourcedir}/%{name}/loadwatch %{buildroot}/%{bin}
+install -m 755 %{_sourcedir}/%{name}/loadwatch.env %{buildroot}/%{conf}
+install -m 0600 %{_sourcedir}/%{name}/loadwatch.cron %{buildroot}/%{cron}
 touch %{buildroot}/etc/plbakeloadwatchinstalled
 
 %pre
-[[ -d /var/log/loadwatch ]] && mkdir /var/log/loadwatch
-[[ -f /root/loadwatch/checklog ]] && mv /root/loadwatch/checklog /var/log/loadwatch/check.log
-[[ -f /var/log/loadwatch.log ]] && mv /var/log/loadwatch.log /var/log/loadwatch/check.log
-[[ -d /root/loadwatch ]] && rsync -aHl /root/loadwatch /var/log/loadwatch >/dev/null
-rm -rf /root/loadwatch
+[[ -d /%{log_dir} ]] && mkdir -p /%{log_dir}
+[[ -f /%{legacy_log_dir}/checklog ]] && mv /%{legacy_log_dir}/checklog /%{log_dir}/check.log
+[[ -f /var/log/loadwatch.log ]] && mv /var/log/loadwatch.log /%{log_dir}/check.log
+if [[ -d /%{legacy_log_dir} ]]; then
+  rsync -aHl /%{legacy_log_dir}/ /%{log_dir}/ >/dev/null
+  rm -rf /root/loadwatch
+fi
 rm -f /root/bin/loadwatch.sh /root/bin/loadwatch
 sed -i -e '/\/root\/bin\/loadwatch/d' -e '/\/root\/loadwatch/d' /var/spool/cron/root
 
+%post
+# with next version uncomment the following line
+# if [[ $1 -eq '1' ]]; then
+  # disable apache statistics if not on cPanel - they can manually be enabled later
+  if [[ -f /usr/local/cpanel/version ]]; then
+    sed -i -e '/^APACHEURI/d' -e '/^APACHEPORT/d' /%{conf}
+
+    port=$(netstat -tpln | \
+      awk '$7 ~ /httpd$/ && $4 ~/[[:digit:]]:.*0$/ {gsub("^.*:", "", $4); print $4}')
+
+    echo "APACHEURI='/whm-server-status'" >> /%{conf}
+    echo "APACHEPORT=${port}" >> /%{conf}
+  fi
+# with the next version, uncomment the next line
+# fi
 
 %clean
 rm -rf ${RPM_BUILD_ROOT}
@@ -53,14 +74,20 @@ rm -rf ${RPM_BUILD_ROOT}
 %files
 %defattr(-,root,root)
 
-%dir /var/log/loadwatch
-%config(noreplace) /etc/default/loadwatch
-/usr/local/lp/bin/loadwatch
-/etc/cron.d/loadwatch.cron
-/root/loadwatch
+%dir /%{log_dir}
+%config(noreplace) /%{conf}
+/%{bin}
+/%{cron}
+/%{legacy_log_dir}
 /etc/plbakeloadwatchinstalled
 
 %changelog
+* Tue May 01 2018 Jack Hayhurst <jhayhurst@liquidweb.com> 1.2.0
+- changed default action to add support for platforms other than cPanel
+- Plesk is now supported
+- added support for alternate Apache Status URI's
+- fixed /var/log/loadwatc/loadwatch bug
+
 * Mon Oct 30 2017 Jack Hayhurst <jhayhurst@liquidweb.com> 1.1.0
 - changed checklog name to /var/log/loadwatch/check.log
 - changed output files to /var/log/loadwatch/date.txt
